requirements.txt
Cython==0.29.24
easydict==1.9
Flask==2.0.1
Flask-Uploads==0.2.1
Flask-WTF==0.15.1
imageio==2.9.0
insightface==0.4.1
Jinja2==3.0.1
mxnet==1.8.0.post0
numpy==1.21.0
onnx==1.9.0
onnxruntime==1.8.0
opencv-python==4.5.2.54
opencv-python-headless==4.5.3.56
openpyxl==3.0.7
Pillow==8.3.0
scikit-image==0.18.2
scikit-learn==0.24.2
scipy==1.7.0
torch==1.9.0
torchaudio==0.9.0
torchvision==0.10.0
tqdm==4.61.1
urllib3==1.26.6
Werkzeug==2.0.1
WTForms==2.3.3
XlsxWriter==1.4.4

SQLAlchemy~=1.3.12
itsdangerous~=2.0.1